Move xdm.variable to xdm.optree.variable.
