+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/full/text/FTScoreVar.kt
@@ -16,7 +16,7 @@
 package uk.co.reecedunn.intellij.plugin.xpath.ast.full.text
 
 import com.intellij.psi.PsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * An XQuery Full Text 1.0 `FTScoreVar` node in the XQuery AST.
