+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/plugin/PluginQuantifiedExprBinding.kt
@@ -16,7 +16,7 @@
 package uk.co.reecedunn.intellij.plugin.xpath.ast.plugin
 
 import com.intellij.psi.PsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * A `QuantifiedExprBinding` node in the XQuery AST.
