+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/xpath/XPathParam.kt
@@ -16,8 +16,8 @@
 package uk.co.reecedunn.intellij.plugin.xpath.ast.xpath
 
 import com.intellij.psi.NavigatablePsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableType
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableType
 
 /**
  * An XPath 3.0 and XQuery 1.0 `Param` node in the XQuery AST.
