+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/xpath/XPathParamList.kt
@@ -17,7 +17,7 @@ package uk.co.reecedunn.intellij.plugin.xpath.ast.xpath
 
 import com.intellij.psi.NavigatablePsiElement
 import com.intellij.util.Range
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * An XPath 3.0 and XQuery 1.0 `ParamList` node in the XQuery AST.
