+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/ast/xpath/XPathSimpleForBinding.kt
@@ -16,7 +16,7 @@
 package uk.co.reecedunn.intellij.plugin.xpath.ast.xpath
 
 import com.intellij.psi.PsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * An XPath 3.0 `SimpleForBinding` node in the XPath AST.
