+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/completion/lookup/XPathVarNameLookup.kt
@@ -18,8 +18,8 @@ package uk.co.reecedunn.intellij.plugin.xpath.completion.lookup
 import com.intellij.psi.PsiElement
 import uk.co.reecedunn.intellij.plugin.xpath.intellij.resources.XPathIcons
 import uk.co.reecedunn.intellij.plugin.xdm.types.element
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableDefinition
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableType
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableDefinition
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableType
 
 class XPathVarNameLookup(localName: String, prefix: String?, private val variable: XpmVariableDefinition) :
     XPathLookupElement(prefix?.let { "$it:$localName" } ?: localName) {
