+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/optree/function/XpmFunctionDeclaration.kt
@@ -20,7 +20,7 @@ import com.intellij.util.Range
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmAnnotation
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmSequenceType
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 interface XpmFunctionDeclaration {
     companion object {
