+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/optree/variable/XpmVariable.kt
@@ -13,6 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.co.reecedunn.intellij.plugin.xpm.variable
+package uk.co.reecedunn.intellij.plugin.xpm.optree.variable
 
-interface XpmVariableReference : XpmVariable
+import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
+
+interface XpmVariable {
+    val variableName: XsQNameValue?
+}
