+++ src/lang-xpm/main/uk/co/reecedunn/intellij/plugin/xpm/optree/variable/XpmVariableType.kt
@@ -13,6 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.co.reecedunn.intellij.plugin.xpm.variable
+package uk.co.reecedunn.intellij.plugin.xpm.optree.variable
 
-interface XpmVariableBinding : XpmVariableDefinition
+import uk.co.reecedunn.intellij.plugin.xdm.types.XdmSequenceType
+
+interface XpmVariableType {
+    val variableType: XdmSequenceType?
+}
