+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/ast/plugin/PluginDefaultCaseClause.kt
@@ -16,7 +16,7 @@
 package uk.co.reecedunn.intellij.plugin.xquery.ast.plugin
 
 import com.intellij.psi.PsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * A `DefaultCaseClause` node in the XQuery AST.
