+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/ast/scripting/ScriptingAssignmentExpr.kt
@@ -17,7 +17,7 @@ package uk.co.reecedunn.intellij.plugin.xquery.ast.scripting
 
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathExprSingle
 import uk.co.reecedunn.intellij.plugin.xpm.optree.XpmExpression
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * An XQuery Scripting Extension 1.0 `AssignmentExpr` node in the XQuery AST.
