+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/ast/xquery/XQuerySlidingWindowClause.kt
@@ -16,7 +16,7 @@
 package uk.co.reecedunn.intellij.plugin.xquery.ast.xquery
 
 import com.intellij.psi.PsiElement
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 
 /**
  * An XQuery 3.0 `SlidingWindowClause` node in the XQuery AST.
