+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/completion/providers/XQueryVarRefProvider.kt
@@ -24,7 +24,7 @@ import uk.co.reecedunn.intellij.plugin.xpath.completion.lookup.XPathVarNameLooku
 import uk.co.reecedunn.intellij.plugin.xpath.completion.property.XPathCompletionProperty
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.EQNameCompletionType
 import uk.co.reecedunn.intellij.plugin.xpath.completion.providers.completionType
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableBinding
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariableBinding
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
 import uk.co.reecedunn.intellij.plugin.xpath.model.inScopeVariables
 import uk.co.reecedunn.intellij.plugin.xpm.context.expand
