+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/psi/impl/xquery/XQueryQNamePsiImpl.kt
@@ -21,7 +21,7 @@ import com.intellij.psi.NavigatablePsiElement
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReference
 import uk.co.reecedunn.intellij.plugin.xpm.optree.function.XpmFunctionReference
-import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariable
+import uk.co.reecedunn.intellij.plugin.xpm.optree.variable.XpmVariable
 import uk.co.reecedunn.intellij.plugin.xpath.psi.impl.xpath.XPathQNamePsiImpl
 import uk.co.reecedunn.intellij.plugin.xquery.psi.impl.reference.XQueryQNamePrefixReference
 import uk.co.reecedunn.intellij.plugin.xpath.psi.impl.reference.XPathFunctionNameReference
